@@ -1,6 +1,6 @@
 {
   "name": "metromed-uc-app",
-  "version": "0.1.0-alpha.1",
+  "version": "0.1.0-alpha.2",
   "description": "Urgent Care app project that uses material-ui",
   "private": true,
   "repository": {
@@ -8,15 +8,18 @@
     "url": "https://github.com/aidarbuy/metromed-ui.git"
   },
   "scripts": {
-    "start": "gulp"
+    "start": "gulp",
+    "dev": "webpack-dev-server --inline --hot"
   },
   "dependencies": {
     "browser-sync-spa": "^1.0.3",
     "express.io": "^1.1.13",
     "google-maps": "^3.2.1",
     "history": "^1.17.0",
+    "jquery": "^2.2.3",
     "marked": "^0.3.5",
     "material-ui": "^0.14.4",
+    "mustache": "^2.2.1",
     "react": "^0.14.8",
     "react-addons-css-transition-group": "^0.14.8",
     "react-dom": "^0.14.8",
@@ -28,23 +31,39 @@
     "socket.io": "^1.4.5"
   },
   "devDependencies": {
+    "babel-core": "^6.7.7",
+    "babel-eslint": "^6.0.3",
+    "babel-loader": "^6.2.4",
     "babel-plugin-transform-object-rest-spread": "^6.6.5",
-    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.3.13",
     "babel-preset-stage-0": "^6.5.0",
     "babelify": "^7.2.0",
+    "baggage-loader": "^0.2.4",
     "browser-sync": "^2.11.0",
     "browserify": "^13.0.0",
+    "clean-webpack-plugin": "^0.1.8",
+    "css-loader": "^0.23.1",
+    "eslint": "^2.8.0",
+    "eslint-loader": "^1.3.0",
+    "extract-text-webpack-plugin": "^1.0.1",
+    "file-loader": "^0.8.5",
     "gulp": "^3.8.10",
     "gulp-notify": "^2.1.0",
+    "gulp-sass": "^2.3.1",
     "gulp-util": "^3.0.1",
+    "html-loader": "^0.4.3",
+    "node-sass": "^3.5.3",
     "pretty-hrtime": "^1.0.1",
     "redux-devtools": "^3.1.1",
     "require-dir": "^0.3.0",
+    "sass-loader": "^3.2.0",
+    "style-loader": "^0.13.1",
     "underscore": "^1.7.0",
+    "url-loader": "^0.5.7",
     "vinyl-source-stream": "^1.0.0",
     "watchify": "^3.6.1",
-    "webpack": "^1.12.14",
+    "webpack": "^1.13.0",
     "webpack-dev-server": "^1.14.1"
   }
 }
