@@ -1,32 +1,26 @@
-import React from 
-  'react'
-import Card from 
-  'material-ui/lib/card/card'
-import CardActions from 
-  'material-ui/lib/card/card-actions'
-import CardHeader from 
-  'material-ui/lib/card/card-header'
-import CardMedia from 
-  'material-ui/lib/card/card-media'
-import CardTitle from 
-  'material-ui/lib/card/card-title'
-import FlatButton from 
-  'material-ui/lib/flat-button'
-import CardText from 
-  'material-ui/lib/card/card-text'
-
-import Conditions from 
-  './Conditions'
+import React       from 'react';
+import Card        from 'material-ui/lib/card/card';
+import CardActions from 'material-ui/lib/card/card-actions';
+import CardHeader  from 'material-ui/lib/card/card-header';
+import CardMedia   from 'material-ui/lib/card/card-media';
+import CardTitle   from 'material-ui/lib/card/card-title';
+import FlatButton  from 'material-ui/lib/flat-button';
+import CardText    from 'material-ui/lib/card/card-text';
+import Conditions  from './Conditions';
 
 module.exports = () => (
   <Card>
     <CardTitle 
-      title={<h4 className="secondary">
-        Welcome to MetroMed!
-      </h4>}
-      subtitle={<h5>
-        952 Edwards Ferry Rd NE, Leesburg, VA 20176
-      </h5>}
+      title={
+        <h4 className="secondary">
+          Welcome to MetroMed!
+        </h4>
+      }
+      subtitle={
+        <h5>
+          952 Edwards Ferry Rd NE, Leesburg, VA 20176
+        </h5>
+      }
     />
 
     <CardMedia overlay={
