@@ -1,38 +1,21 @@
-import React from 
-  'react';
-import StarBorder from 
-  'material-ui/lib/svg-icons/toggle/star-border'
-import Toolbar from 
-  'material-ui/lib/toolbar/toolbar'
-import ToolbarGroup from 
-  'material-ui/lib/toolbar/toolbar-group'
-import ToolbarSeparator from 
-  'material-ui/lib/toolbar/toolbar-separator'
-import ToolbarTitle from 
-  'material-ui/lib/toolbar/toolbar-title'
-import DropDownMenu from 
-  'material-ui/lib/DropDownMenu'
-import IconMenu from 
-  'material-ui/lib/menus/icon-menu'
-import IconButton from 
-  'material-ui/lib/icon-button'
-import FontIcon from 
-  'material-ui/lib/font-icon'
-import NavigationExpandMoreIcon from 
-  'material-ui/lib/svg-icons/navigation/expand-more'
-import MenuItem from 
-  'material-ui/lib/menus/menu-item'
-import RaisedButton from 
-  'material-ui/lib/raised-button'
-import Paper from 
-  'material-ui/lib/paper'
-import { grey50 } from
-  'material-ui/lib/styles/colors'
+import React from 'react';
+import StarBorder from 'material-ui/lib/svg-icons/toggle/star-border';
+import Toolbar from 'material-ui/lib/toolbar/toolbar';
+import ToolbarGroup from 'material-ui/lib/toolbar/toolbar-group';
+import ToolbarSeparator from 'material-ui/lib/toolbar/toolbar-separator';
+import ToolbarTitle from 'material-ui/lib/toolbar/toolbar-title';
+import DropDownMenu from 'material-ui/lib/DropDownMenu';
+import IconMenu from 'material-ui/lib/menus/icon-menu';
+import IconButton from 'material-ui/lib/icon-button';
+import FontIcon from 'material-ui/lib/font-icon';
+import NavigationExpandMoreIcon from 'material-ui/lib/svg-icons/navigation/expand-more';
+import MenuItem from 'material-ui/lib/menus/menu-item';
+import RaisedButton from 'material-ui/lib/raised-button';
+import Paper from 'material-ui/lib/paper';
+import { grey50 } from'material-ui/lib/styles/colors';
 
-import articlesData from 
-  '../../data/articles'
-import ArticleCard from 
-  '../ArticleCard';
+import articlesData from '../../data/articles';
+import ArticleCard from '../ArticleCard';
 
 const articles = articlesData.slice(0, 4);
 
@@ -66,13 +49,21 @@ module.exports = () => (
     </div>
 
     <Toolbar style={{background:grey50}}>
-      <ToolbarGroup style={{width:'100%'}}
+      <ToolbarGroup style={{
+        width:'100%'
+      }}
         firstChild={true}
         lastChild={true}
       >
         <RaisedButton 
           fullWidth={true}
-          label="View All Articles" 
+          style={{
+            textAlign:'center',
+            // margin:0,
+            // border:'1px dashed green',
+            // width:'110%',
+          }}
+          label="View All Articles"
           linkButton={true}
           href="/articles"
         />
