@@ -1,42 +1,41 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
+import React from 'react';
 
 const styles = {
   main: {
     marginTop: 20,
-    width: 1160, 
-    height: 600, 
+    width: 1160,
+    height: 600,
   },
   bgBlur: {
-    top: '90%', 
+    top: '90%',
     left: '50%',
   },
   getStyles: () => {
     return {
-      top: '89%', 
-      fontWeight: 300, 
-      fontSize: '60', 
-      color: 'rgb(86, 204, 225)', 
+      top: '89%',
+      fontWeight: 300,
+      fontSize: '60',
+      color: 'rgb(86, 204, 225)',
       whiteSpace: 'nowrap',
-      left: '10%'
-    }
+      left: '10%',
+    };
   },
   getPlus: () => {
     return {
-      left: 50, 
-      fontWeight: 'normal', 
-      textAlign: 'center', 
-      width: 30, 
-      height: 30, 
-      fontSize: 30, 
-      lineHeight: '30px', 
-      color: 'white', 
-      background: '#2196F3', 
-      borderRadius: 100, 
-      whiteSpace: 'nowrap'
+      left: 50,
+      fontWeight: 'normal',
+      textAlign: 'center',
+      width: 30,
+      height: 30,
+      fontSize: 30,
+      lineHeight: '30px',
+      color: 'white',
+      background: '#2196F3',
+      borderRadius: 100,
+      whiteSpace: 'nowrap',
     };
-  }
-}
+  },
+};
 
 styles['slideA']          = {we: styles.getStyles()}
 styles.slideA['care']     = styles.getStyles()
@@ -203,38 +202,89 @@ const LayerSlider = React.createClass({
           <p className="ls-l" style={styles.slideA.hours} data-ls="delayin:17500; offsetxin:0; offsetyin:10; durationin:2000; showuntil:3000; offsetxout:0; offsetyout:0; durationout:2000;">Monday &ndash; Friday: 10 am &ndash; 8 pm</p>
         </div>
 
+
         {/* 11000 30-32, 69-72 */}
-        <div className="ls-slide" data-ls="slidedelay:11000; transition2d: 15-18, 59-61, 90, 91, 30-32, 69-72, 40-43, 92, 96, 97, 75-79, 81, 82, 86, 87-89, 93-95, 111, 103-105, 109, 112">
+        <div className="ls-slide" data-ls="slidedelay:11000;
+          transition2d: 15-18, 59-61, 90, 91, 30-32, 69-72, 40-43, 92, 96,
+          97, 75-79, 81, 82, 86, 87-89, 93-95, 111, 103-105, 109, 112"
+        >
           <img className="ls-bg" src="images/layerslider/image-2.jpg" alt="Slide background" />
-          <p className="ls-l" style={styles.slideB.fullSpectrum} data-ls="offsetxin:0; durationin:2000; delayin:500; easingin:easeOutElastic; rotatexin:-90; transformoriginin:50% top 0; offsetxout:-200; durationout:1000;">Full Spectrum</p>
-          <p className="ls-l" style={styles.slideB.treatment} data-ls="  offsetxin:0; durationin:2000; delayin:1000; easingin:easeOutElastic; rotatexin:90; transformoriginin:50% top 0; offsetxout:-400;">t r e a t m e n t</p>
+          <p className="ls-l" style={styles.slideB.fullSpectrum}
+            data-ls="offsetxin:0; durationin:2000; delayin:500; easingin:easeOutElastic;
+              rotatexin:-90; transformoriginin:50% top 0; offsetxout:-200; durationout:1000;"
+          >Full Spectrum</p>
+
+          <p className="ls-l" style={styles.slideB.treatment}
+            data-ls="offsetxin:0; durationin:2000; delayin:1000; easingin:easeOutElastic;
+              rotatexin:90; transformoriginin:50% top 0; offsetxout:-400;"
+          >t r e a t m e n t</p>
+
           <h5 className="ls-l" style={styles.slideB.counseling} data-ls="offsetxin:0; delayin:2000; easingin:easeOutQuint; scalexin:0.8; scaleyin:0.8; offsetxout:0; durationout:750; scalexout:0.8; scaleyout:0.8;">Medical Counseling</h5>
+
           <h5 className="ls-l" style={styles.slideB.clinic} data-ls="offsetxin:0; delayin:2500; easingin:easeOutQuint; scalexin:0.8; scaleyin:0.8; offsetxout:0; durationout:750; scalexout:0.8; scaleyout:0.8;">Cardiac Clinic</h5>
+
           <h5 className="ls-l" style={styles.slideB.pediatric} data-ls="offsetxin:0; delayin:3000; easingin:easeOutQuint; scalexin:0.8; scaleyin:0.8; offsetxout:0; durationout:750; scalexout:0.8; scaleyout:0.8;">Pediatric Clinic</h5>
+
           <h5 className="ls-l" style={styles.slideB.gynecological} data-ls="offsetxin:0; delayin:3500; easingin:easeOutQuint; scalexin:0.8; scaleyin:0.8; offsetxout:0; durationout:750; scalexout:0.8; scaleyout:0.8;">Gynecological Clinic</h5>
+
           <h5 className="ls-l" style={styles.slideB.lab} data-ls="offsetxin:0; delayin:4000; easingin:easeOutQuint; scalexin:0.8; scaleyin:0.8; offsetxout:0; durationout:750; scalexout:0.8; scaleyout:0.8;">Laboratory Analysis</h5>
+
           <h5 className="ls-l" style={styles.slideB.diag} data-ls="offsetxin:0; delayin:4500; easingin:easeOutQuint; scalexin:0.8; scaleyin:0.8; offsetxout:0; durationout:750; scalexout:0.8; scaleyout:0.8;">Diagnosis Clinic</h5>
+
           <h5 className="ls-l" style={styles.slideB.pluses[0]} data-ls="offsetxin:0; durationin:750; delayin:1800; easingin:easeOutQuint; rotatein:90; scalexin:0.5; scaleyin:0.5; offsetxout:0; durationout:750; rotateout:90; scalexout:0.5; scaleyout:0.5;">+</h5>
+
           <h5 className="ls-l" style={styles.slideB.pluses[1]} data-ls="offsetxin:0; durationin:750; delayin:2300; easingin:easeOutQuint; rotatein:90; scalexin:0.5; scaleyin:0.5; offsetxout:0; durationout:750; rotateout:90; scalexout:0.5; scaleyout:0.5;">+</h5>
+
           <h5 className="ls-l" style={styles.slideB.pluses[2]} data-ls="offsetxin:0; durationin:750; delayin:2800; easingin:easeOutQuint; rotatein:90; scalexin:0.5; scaleyin:0.5; offsetxout:0; durationout:750; rotateout:90; scalexout:0.5; scaleyout:0.5;">+</h5>
+
           <h5 className="ls-l" style={styles.slideB.pluses[3]} data-ls="offsetxin:0; durationin:750; delayin:3300; easingin:easeOutQuint; rotatein:90; scalexin:0.5; scaleyin:0.5; offsetxout:0; durationout:750; rotateout:90; scalexout:0.5; scaleyout:0.5;">+</h5>
+
           <h5 className="ls-l" style={styles.slideB.pluses[4]} data-ls="offsetxin:0; durationin:750; delayin:3800; easingin:easeOutQuint; rotatein:90; scalexin:0.5; scaleyin:0.5; offsetxout:0; durationout:750; rotateout:90; scalexout:0.5; scaleyout:0.5;">+</h5>
+
           <h5 className="ls-l" style={styles.slideB.pluses[5]} data-ls="offsetxin:0; durationin:750; delayin:4300; easingin:easeOutQuint; rotatein:90; scalexin:0.5; scaleyin:0.5; offsetxout:0; durationout:750; rotateout:90; scalexout:0.5; scaleyout:0.5;">+</h5>
         </div>
 
+
         {/* 10500 40-43, 92, 96, 97 */}
-        <div className="ls-slide" data-ls="slidedelay:10500; transition2d: 15-18, 59-61, 90, 91, 30-32, 69-72, 40-43, 92, 96, 97, 75-79, 81, 82, 86, 87-89, 93-95, 111, 103-105, 109, 112">
+        <div className="ls-slide" data-ls="slidedelay:10500;
+          transition2d: 15-18, 59-61, 90, 91, 30-32, 69-72, 40-43, 92, 96,
+            97, 75-79, 81, 82, 86, 87-89, 93-95, 111, 103-105, 109, 112"
+        >
           <img className="ls-bg" src="images/layerslider/image-3.jpg" alt="Slide background" />
-          <h4 className="ls-l" style={{top: 100, left: 125, whiteSpace: 'nowrap', fontSize: 70, fontWeight: 'normal', color: '#F50057'}} data-ls="offsetxin:0; offsetyin:top; durationin:3000; delayin:300; easingin:easeOutElastic; fadein:false; offsetxout:left; durationout:1500; showuntil:5400; easingout:easeInBack; fadeout:false;">Playful</h4>
-          <h3 className="ls-l" style={{top: 175, left: 120, whiteSpace: 'nowrap', fontSize: 100, fontWeight: 'normal', color: '#FF5722'}} data-ls="offsetxin:0; offsetyin:top; durationin:3000; delayin:200; easingin:easeOutElastic; fadein:false; offsetxout:left; durationout:1500; showuntil:5950; easingout:easeInBack; fadeout:false;">Pediatric</h3>
-          <h4 className="ls-l" style={{top: 290, left: 390, whiteSpace: 'nowrap', fontSize: 70, fontWeight: 'normal', color: '#F50057'}} data-ls="offsetxin:0; offsetyin:top; durationin:3000; delayin:100; easingin:easeOutElastic; fadein:false; offsetxout:left; durationout:1500; showuntil:5250; easingout:easeInBack; fadeout:false;">Care</h4>
-          <h5 className="ls-l" style={{top: 450, left: 125, fontSize: 25, color: '#FF3D00'}} data-ls="offsetxin:-100; offsetxout:0; offsetyin:0; offsetyout:bottom; durationin:2500; delayin:2000; skewxin:60;">When medical treatment is fun :)</h5>
+
+          <h4 className="ls-l" style={{ top: 100, left: 125, whiteSpace: 'nowrap',
+            fontSize: 70, fontWeight: 'normal', color: '#F50057' }}
+            data-ls="offsetxin:0; offsetyin:top; durationin:3000; delayin:300;
+            easingin:easeOutElastic; fadein:false; offsetxout:left; durationout:1500;
+            showuntil:5400; easingout:easeInBack; fadeout:false;"
+          >Playful</h4>
+
+          <h3 className="ls-l" style={{ top: 175, left: 120, whiteSpace: 'nowrap',
+            fontSize: 100, fontWeight: 'normal', color: '#FF5722' }}
+            data-ls="offsetxin:0; offsetyin:top; durationin:3000; delayin:200;
+              easingin:easeOutElastic; fadein:false; offsetxout:left;
+              durationout:1500; showuntil:5950; easingout:easeInBack; fadeout:false;"
+          >Pediatric</h3>
+
+          <h4 className="ls-l"
+            style={{ top: 290, left: 390, whiteSpace: 'nowrap',
+              fontSize: 70, fontWeight: 'normal', color: '#F50057' }}
+            data-ls="offsetxin:0; offsetyin:top; durationin:3000; delayin:100;
+              easingin:easeOutElastic; fadein:false; offsetxout:left; durationout:1500;
+              showuntil:5250; easingout:easeInBack; fadeout:false;"
+          >Care</h4>
+
+          <h5 className="ls-l" style={{ top: 450, left: 125, fontSize: 25, color: '#FF3D00' }}
+            data-ls="offsetxin:-100; offsetxout:0; offsetyin:0; offsetyout:bottom;
+              durationin:2500; delayin:2000; skewxin:60;"
+          >When medical treatment is fun :)</h5>
+
           {/* 75-79, 81, 82, 86, 87-89, 93-95, 111, 103-105, 109, 112
           3d: 19, 20, 22, 23, 26, 62 63, 68, 69, 76 */}
         </div>
       </div>
-    )
-  }
-})
+    );
+  },
+});
 
 export default LayerSlider;
